feat: add support for wasm32 target
This commit enables compilation for the `wasm32-unknown-unknown` target,
allowing the `binwalk` library to be used in browser environments.

Key changes:
- Gated non-portable dependencies (`termsize`, `walkdir`, `threadpool`, `delink`) in `Cargo.toml`.
- Applied `#![cfg(not(target_arch = "wasm32"))]` to `src/main.rs` to keep the CLI desktop-only.
- Conditionally compiled or stubbed filesystem and process logic in `src/extractors/common.rs`, `src/binwalk.rs`, and `src/common.rs`.
- Preserved portable internal extractors for high-confidence signature validation on WASM.
- Fixed 32-bit arithmetic overflows and relaxed 64-bit-only checks in structural parsers.
- Added `README.wasm.md` summarizing the WASM-specific build process and limitations.

+++ src/structures/lzfse.rs
@@ -102,9 +102,9 @@ fn parse_compressedv1_block_header(lzfse_data: &[u8]) -> Result<LZFSEBlock, Stru
 
 /// Parse a compressed (version 2) LZFSE block header
 fn parse_compressedv2_block_header(lzfse_data: &[u8]) -> Result<LZFSEBlock, StructureError> {
-    const N_PAYLOAD_SHIFT: usize = 20;
-    const LMD_PAYLOAD_SHIFT: usize = 40;
-    const PAYLOAD_MASK: usize = 0b11111_11111_11111_11111;
+    const N_PAYLOAD_SHIFT: u32 = 20;
+    const LMD_PAYLOAD_SHIFT: u32 = 40;
+    const PAYLOAD_MASK: u64 = 0b11111_11111_11111_11111;
 
     let block_structure = vec![
         ("magic", "u32"),
@@ -118,13 +118,13 @@ fn parse_compressedv2_block_header(lzfse_data: &[u8]) -> Result<LZFSEBlock, Stru
 
     if let Ok(block_header) = parse(lzfse_data, &block_structure, "little") {
         let n_lmd_payload_bytes =
-            (block_header["packed_field_2"] >> LMD_PAYLOAD_SHIFT) & PAYLOAD_MASK;
+            ((block_header["packed_field_2"] as u64) >> LMD_PAYLOAD_SHIFT) & PAYLOAD_MASK;
         let n_literal_payload_bytes =
-            (block_header["packed_field_1"] >> N_PAYLOAD_SHIFT) & PAYLOAD_MASK;
+            ((block_header["packed_field_1"] as u64) >> N_PAYLOAD_SHIFT) & PAYLOAD_MASK;
 
         return Ok(LZFSEBlock {
             eof: false,
-            data_size: n_lmd_payload_bytes + n_literal_payload_bytes,
+            data_size: (n_lmd_payload_bytes + n_literal_payload_bytes) as usize,
             header_size: block_header["header_size"],
         });
     }
